#238 refactor: battle enum 패키지 이동
