+++ src/main/java/UMC_7th/Closit/domain/battle/entity/Battle.java
@@ -1,5 +1,6 @@
 package UMC_7th.Closit.domain.battle.entity;
 
+import UMC_7th.Closit.domain.battle.entity.enums.BattleStatus;
 import UMC_7th.Closit.domain.post.entity.Post;
 import UMC_7th.Closit.global.common.BaseEntity;
 import jakarta.persistence.*;
