+++ src/main/java/UMC_7th/Closit/domain/battle/entity/enums/BattleSorting.java
@@ -1,4 +1,4 @@
-package UMC_7th.Closit.domain.battle.entity;
+package UMC_7th.Closit.domain.battle.entity.enums;
 
 import lombok.Getter;
 import org.springframework.data.domain.Sort;
