+++ src/main/java/UMC_7th/Closit/domain/battle/entity/enums/BattleStatus.java
@@ -1,4 +1,4 @@
-package UMC_7th.Closit.domain.battle.entity;
+package UMC_7th.Closit.domain.battle.entity.enums;
 
 public enum BattleStatus {
     INACTIVE, PENDING, ACTIVE, COMPLETED
