+++ src/main/java/UMC_7th/Closit/domain/battle/entity/enums/ChallengeStatus.java
@@ -1,4 +1,4 @@
-package UMC_7th.Closit.domain.battle.entity;
+package UMC_7th.Closit.domain.battle.entity.enums;
 
 public enum ChallengeStatus {
     PENDING, ACCEPTED, REJECTED
